Refactor API Keys page by extracting modals and hooks into separate components
