+++ autogpt_platform/frontend/src/app/(platform)/profile/(user)/api_keys/components/APIKeySection/APIKeySection.tsx
@@ -1,0 +1,102 @@
+"use client";
+
+import { Loader2, MoreVertical } from "lucide-react";
+import { Button } from "@/components/ui/button";
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table";
+import { Badge } from "@/components/ui/badge";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+import { useAPISection } from "./useAPISection";
+
+export function APIKeysSection() {
+  const { apiKeys, isLoading, isDeleting, handleRevokeKey } = useAPISection();
+
+  return (
+    <>
+      {isLoading ? (
+        <div className="flex justify-center p-4">
+          <Loader2 className="h-6 w-6 animate-spin" />
+        </div>
+      ) : (
+        apiKeys &&
+        apiKeys.length > 0 && (
+          <Table>
+            <TableHeader>
+              <TableRow>
+                <TableHead>Name</TableHead>
+                <TableHead>API Key</TableHead>
+                <TableHead>Status</TableHead>
+                <TableHead>Created</TableHead>
+                <TableHead>Last Used</TableHead>
+                <TableHead></TableHead>
+              </TableRow>
+            </TableHeader>
+            <TableBody>
+              {apiKeys.map((key) => (
+                <TableRow key={key.id}>
+                  <TableCell>{key.name}</TableCell>
+                  <TableCell>
+                    <div className="rounded-md border p-1 px-2 text-xs">
+                      {`${key.prefix}******************${key.postfix}`}
+                    </div>
+                  </TableCell>
+                  <TableCell>
+                    <Badge
+                      variant={
+                        key.status === "ACTIVE" ? "default" : "destructive"
+                      }
+                      className={
+                        key.status === "ACTIVE"
+                          ? "border-green-600 bg-green-100 text-green-800"
+                          : "border-red-600 bg-red-100 text-red-800"
+                      }
+                    >
+                      {key.status}
+                    </Badge>
+                  </TableCell>
+                  <TableCell>
+                    {new Date(key.created_at).toLocaleDateString()}
+                  </TableCell>
+                  <TableCell>
+                    {key.last_used_at
+                      ? new Date(key.last_used_at).toLocaleDateString()
+                      : "Never"}
+                  </TableCell>
+                  <TableCell>
+                    <DropdownMenu>
+                      <DropdownMenuTrigger asChild>
+                        <Button variant="ghost" size="sm">
+                          <MoreVertical className="h-4 w-4" />
+                        </Button>
+                      </DropdownMenuTrigger>
+                      <DropdownMenuContent align="end">
+                        <DropdownMenuItem
+                          className="text-destructive"
+                          onClick={() => handleRevokeKey(key.id)}
+                          disabled={isDeleting}
+                        >
+                          Revoke
+                        </DropdownMenuItem>
+                      </DropdownMenuContent>
+                    </DropdownMenu>
+                  </TableCell>
+                </TableRow>
+              ))}
+            </TableBody>
+          </Table>
+        )
+      )}
+    </>
+  );
+}
