+++ autogpt_platform/frontend/src/app/(platform)/profile/(user)/api_keys/components/APIKeySection/useAPISection.tsx
@@ -1,0 +1,61 @@
+"use client";
+import {
+  getGetV1ListUserApiKeysQueryKey,
+  useDeleteV1RevokeApiKey,
+  useGetV1ListUserApiKeys,
+} from "@/api/__generated__/endpoints/api-keys/api-keys";
+import { APIKeyWithoutHash } from "@/api/__generated__/models/aPIKeyWithoutHash";
+import { useToast } from "@/components/ui/use-toast";
+import { getQueryClient } from "@/lib/react-query/queryClient";
+
+export const useAPISection = () => {
+  const queryClient = getQueryClient();
+  const { toast } = useToast();
+
+  const { data: apiKeys, isLoading } = useGetV1ListUserApiKeys({
+    query: {
+      select: (res) => {
+        return (res.data as APIKeyWithoutHash[]).filter(
+          (key) => key.status === "ACTIVE",
+        );
+      },
+    },
+  });
+
+  const { mutateAsync: revokeAPIKey, isPending: isDeleting } =
+    useDeleteV1RevokeApiKey({
+      mutation: {
+        onSettled: () => {
+          return queryClient.invalidateQueries({
+            queryKey: getGetV1ListUserApiKeysQueryKey(),
+          });
+        },
+      },
+    });
+
+  const handleRevokeKey = async (keyId: string) => {
+    try {
+      await revokeAPIKey({
+        keyId: keyId,
+      });
+
+      toast({
+        title: "Success",
+        description: "AutoGPT Platform API key revoked successfully",
+      });
+    } catch {
+      toast({
+        title: "Error",
+        description: "Failed to revoke AutoGPT Platform API key",
+        variant: "destructive",
+      });
+    }
+  };
+
+  return {
+    apiKeys,
+    isLoading,
+    isDeleting,
+    handleRevokeKey,
+  };
+};
