+++ autogpt_platform/frontend/src/app/(platform)/profile/(user)/settings/components/SettingsForm/helper.ts
