+++ autogpt_platform/frontend/src/app/(platform)/profile/(user)/settings/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useGetV1GetNotificationPreferences } from "@/api/__generated__/endpoints/auth/auth";
-import { SettingsForm } from "@/app/(platform)/profile/(user)/settings/components/SettingsForm";
+import { SettingsForm } from "@/app/(platform)/profile/(user)/settings/components/SettingsForm/SettingsForm";
 import { useSupabase } from "@/lib/supabase/hooks/useSupabase";
 import * as React from "react";
 import SettingsLoading from "./loading";
